pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monitorx"
version = "0.1.0"
description = "ML/AI Infrastructure Observability Platform"
authors = [{name = "MonitorX Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
keywords = ["ml", "ai", "monitoring", "observability", "infrastructure"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "influxdb-client>=1.38.0",
    "pydantic>=2.4.2",
    "streamlit>=1.28.1",
    "pandas>=2.1.3",
    "numpy>=1.25.2",
    "plotly>=5.17.0",
    "httpx>=0.25.2",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "APScheduler>=3.10.4",
    "websockets>=12.0",
    "aiofiles>=23.2.1",
    "jinja2>=3.1.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.9.1",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.1",
    "pre-commit>=3.5.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true